Calculate the absolute probability of an index case to case interval belonging to a transmission route component

Calculate the absolute probability of an index case to case interval belonging to a transmission route component

Usage

get_abs_prob(dat, sigma = sd(dat), mu = mean(dat))

Arguments

dat

a (numeric) vector of index case to case intervals

sigma

standard deviation of underlying probability distribution; defaults to sd(dat)

mu

mean of underlying probability distribution; defaults to mean(dat)

Value

matrix where each row represents a transmission route and each column represents the absolue probability of belonging to that transmission route for each index case to case interval in dat

Examples

my_data<-c(rep(1,38),rep(2,39),rep(3,30),rep(4,17),rep(5,7))
get_abs_prob(my_data)
#>              [,1]         [,2]         [,3]         [,4]         [,5]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,11]        [,12]        [,13]        [,14]        [,15]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,16]        [,17]        [,18]        [,19]        [,20]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,21]        [,22]        [,23]        [,24]        [,25]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,26]        [,27]        [,28]        [,29]        [,30]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,31]        [,32]        [,33]        [,34]        [,35]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 0.3912638887 0.3912638887
#> [2,] 0.1769686903 0.1769686903 0.1769686903 0.1769686903 0.1769686903
#> [3,] 0.0086922187 0.0086922187 0.0086922187 0.0086922187 0.0086922187
#> [4,] 0.0225873861 0.0225873861 0.0225873861 0.0225873861 0.0225873861
#> [5,] 0.0009365418 0.0009365418 0.0009365418 0.0009365418 0.0009365418
#> [6,] 0.0027897508 0.0027897508 0.0027897508 0.0027897508 0.0027897508
#> [7,] 0.0001089602 0.0001089602 0.0001089602 0.0001089602 0.0001089602
#>             [,36]        [,37]        [,38]        [,39]        [,40]
#> [1,] 0.3912638887 0.3912638887 0.3912638887 2.361524e-01 2.361524e-01
#> [2,] 0.1769686903 0.1769686903 0.1769686903 3.055789e-01 3.055789e-01
#> [3,] 0.0086922187 0.0086922187 0.0086922187 7.282249e-04 7.282249e-04
#> [4,] 0.0225873861 0.0225873861 0.0225873861 6.678008e-02 6.678008e-02
#> [5,] 0.0009365418 0.0009365418 0.0009365418 1.146018e-04 1.146018e-04
#> [6,] 0.0027897508 0.0027897508 0.0027897508 1.000244e-02 1.000244e-02
#> [7,] 0.0001089602 0.0001089602 0.0001089602 1.524998e-05 1.524998e-05
#>             [,41]        [,42]        [,43]        [,44]        [,45]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,46]        [,47]        [,48]        [,49]        [,50]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,51]        [,52]        [,53]        [,54]        [,55]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,56]        [,57]        [,58]        [,59]        [,60]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,61]        [,62]        [,63]        [,64]        [,65]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,66]        [,67]        [,68]        [,69]        [,70]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,71]        [,72]        [,73]        [,74]        [,75]
#> [1,] 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01 2.361524e-01
#> [2,] 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01 3.055789e-01
#> [3,] 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04 7.282249e-04
#> [4,] 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02 6.678008e-02
#> [5,] 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04 1.146018e-04
#> [6,] 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02 1.000244e-02
#> [7,] 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05 1.524998e-05
#>             [,76]        [,77]        [,78]        [,79]        [,80]
#> [1,] 2.361524e-01 2.361524e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 3.055789e-01 3.055789e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 7.282249e-04 7.282249e-04 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 6.678008e-02 6.678008e-02 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 1.146018e-04 1.146018e-04 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 1.000244e-02 1.000244e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.524998e-05 1.524998e-05 1.695094e-06 1.695094e-06 1.695094e-06
#>             [,81]        [,82]        [,83]        [,84]        [,85]
#> [1,] 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06
#>             [,86]        [,87]        [,88]        [,89]        [,90]
#> [1,] 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06
#>             [,91]        [,92]        [,93]        [,94]        [,95]
#> [1,] 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06
#>             [,96]        [,97]        [,98]        [,99]       [,100]
#> [1,] 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06
#>            [,101]       [,102]       [,103]       [,104]       [,105]
#> [1,] 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01 1.017574e-01
#> [2,] 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01 2.793416e-01
#> [3,] 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05 3.157102e-05
#> [4,] 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01 1.409022e-01
#> [5,] 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06 9.972899e-06
#> [6,] 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02 2.851340e-02
#> [7,] 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06 1.695094e-06
#>            [,106]       [,107]       [,108]       [,109]       [,110]
#> [1,] 1.017574e-01 1.017574e-01 3.128631e-02 3.128631e-02 3.128631e-02
#> [2,] 2.793416e-01 2.793416e-01 1.351273e-01 1.351273e-01 1.351273e-01
#> [3,] 3.157102e-05 3.157102e-05 7.026631e-07 7.026631e-07 7.026631e-07
#> [4,] 1.409022e-01 1.409022e-01 2.122710e-01 2.122710e-01 2.122710e-01
#> [5,] 9.972899e-06 9.972899e-06 6.165176e-07 6.165176e-07 6.165176e-07
#> [6,] 2.851340e-02 2.851340e-02 6.463719e-02 6.463719e-02 6.463719e-02
#> [7,] 1.695094e-06 1.695094e-06 1.495897e-07 1.495897e-07 1.495897e-07
#>            [,111]       [,112]       [,113]       [,114]       [,115]
#> [1,] 3.128631e-02 3.128631e-02 3.128631e-02 3.128631e-02 3.128631e-02
#> [2,] 1.351273e-01 1.351273e-01 1.351273e-01 1.351273e-01 1.351273e-01
#> [3,] 7.026631e-07 7.026631e-07 7.026631e-07 7.026631e-07 7.026631e-07
#> [4,] 2.122710e-01 2.122710e-01 2.122710e-01 2.122710e-01 2.122710e-01
#> [5,] 6.165176e-07 6.165176e-07 6.165176e-07 6.165176e-07 6.165176e-07
#> [6,] 6.463719e-02 6.463719e-02 6.463719e-02 6.463719e-02 6.463719e-02
#> [7,] 1.495897e-07 1.495897e-07 1.495897e-07 1.495897e-07 1.495897e-07
#>            [,116]       [,117]       [,118]       [,119]       [,120]
#> [1,] 3.128631e-02 3.128631e-02 3.128631e-02 3.128631e-02 3.128631e-02
#> [2,] 1.351273e-01 1.351273e-01 1.351273e-01 1.351273e-01 1.351273e-01
#> [3,] 7.026631e-07 7.026631e-07 7.026631e-07 7.026631e-07 7.026631e-07
#> [4,] 2.122710e-01 2.122710e-01 2.122710e-01 2.122710e-01 2.122710e-01
#> [5,] 6.165176e-07 6.165176e-07 6.165176e-07 6.165176e-07 6.165176e-07
#> [6,] 6.463719e-02 6.463719e-02 6.463719e-02 6.463719e-02 6.463719e-02
#> [7,] 1.495897e-07 1.495897e-07 1.495897e-07 1.495897e-07 1.495897e-07
#>            [,121]       [,122]       [,123]       [,124]       [,125]
#> [1,] 3.128631e-02 3.128631e-02 3.128631e-02 3.128631e-02 6.858697e-03
#> [2,] 1.351273e-01 1.351273e-01 1.351273e-01 1.351273e-01 3.447183e-02
#> [3,] 7.026631e-07 7.026631e-07 7.026631e-07 7.026631e-07 7.969560e-09
#> [4,] 2.122710e-01 2.122710e-01 2.122710e-01 2.122710e-01 2.283949e-01
#> [5,] 6.165176e-07 6.165176e-07 6.165176e-07 6.165176e-07 2.704492e-08
#> [6,] 6.463719e-02 6.463719e-02 6.463719e-02 6.463719e-02 1.165406e-01
#> [7,] 1.495897e-07 1.495897e-07 1.495897e-07 1.495897e-07 1.047729e-08
#>            [,126]       [,127]       [,128]       [,129]       [,130]
#> [1,] 6.858697e-03 6.858697e-03 6.858697e-03 6.858697e-03 6.858697e-03
#> [2,] 3.447183e-02 3.447183e-02 3.447183e-02 3.447183e-02 3.447183e-02
#> [3,] 7.969560e-09 7.969560e-09 7.969560e-09 7.969560e-09 7.969560e-09
#> [4,] 2.283949e-01 2.283949e-01 2.283949e-01 2.283949e-01 2.283949e-01
#> [5,] 2.704492e-08 2.704492e-08 2.704492e-08 2.704492e-08 2.704492e-08
#> [6,] 1.165406e-01 1.165406e-01 1.165406e-01 1.165406e-01 1.165406e-01
#> [7,] 1.047729e-08 1.047729e-08 1.047729e-08 1.047729e-08 1.047729e-08
#>            [,131]
#> [1,] 6.858697e-03
#> [2,] 3.447183e-02
#> [3,] 7.969560e-09
#> [4,] 2.283949e-01
#> [5,] 2.704492e-08
#> [6,] 1.165406e-01
#> [7,] 1.047729e-08