Function to integrate SEIR

Function to integrate SEIR

Usage

integrate_seir(beta_t, sigma, gamma, N, T_final, E0, I0, dt = 0.01)

Arguments

beta_t

Function that returns beta (t) for any time t.

sigma

Rate of leaving latent class.

gamma

Rate of leaving infected class.

N

Total population size.

T_final

Final time step.

E0

Initial number of individuals in exposed state.

I0

Initial number of individuals in infectious state.

dt

Size of time step over which to integrate SEIR; defaults to dt = 0.01.

Value

Data frame with number of individuals in each state for each time point.