Calculate f0 for Different Components

This function calculates the value of f0 based on the component, where the components represent the transmission routes: Co-Primary (CP), Primary-Secondary (PS), Primary-Tertiary (PT), and Primary-Quaternary (PQ). We split the PS, PT and PQ routes into two parts, such that

component 1: CP route

component 2+3: PS route

component 4+5: PT route

component 6+7: PQ route

Usage

f0(x, mu, sigma, comp, dist = "normal")

Arguments

x

numeric; the value at which to evaluate the function.

mu

numeric; the mean value.

sigma

numeric; the standard deviation.

comp

integer; the component number (1 to 7).

dist

string; assumed distribution of the serial interval; takes "normal" or "gamma"; defaults to "normal"

Value

The calculated value of f0.

Details

If the dist = gamma, then the mean (\(\mu\)) and standard deviation (sigma) are converted into the shape (k) and scale (theta) parameters for the gamma distribution, such that the mean (\(\mu\) ) and variance (\(\sigma^2\)) are given by: $$\mu = k \times \theta$$ $$\sigma^2 = k \times \theta^2$$.